+++ Y2Snes.Core/Cpu/Instruction.cs
@@ -7,9 +7,9 @@ namespace Y2Snes.Core
         public string Name { get; }
         public byte OpCode { get; }
         public byte OperandLength { get; }
-        public Action<ushort> Handler { get; }
+        public Action<uint> Handler { get; }
 
-        public Instruction(string name, byte opCode, byte operandLength, Action<ushort> handler)
+        public Instruction(string name, byte opCode, byte operandLength, Action<uint> handler)
         {
             Name = name;
             OpCode = opCode;
